Relocate token auth middleware
